Fix watch for events; add test for interface implementation so it won't break again.
