+++ pkg/api/errors/etcd/etcd.go
@@ -64,10 +64,3 @@ func InterpretDeleteError(err error, kind, name string) error {
 		return err
 	}
 }
-
-// InterpretResourceVersionError returns the appropriate api error
-// for a failure to convert the resource version of an object sent
-// to the API to an etcd uint64 index.
-func InterpretResourceVersionError(err error, kind, value string) error {
-	return errors.NewInvalid(kind, "", errors.ValidationErrorList{errors.NewFieldInvalid("resourceVersion", value)})
-}
